Moves PieceColor to 'model' package
